Module Eq

Eq plugin: receive another value as inherited attribute and test for equality.

Very similar to Compare plugin.

For type declaration type ('a,'b,...) typ = ... it will create a transformation function with type

('a -> 'a -> bool) -> ('b -> 'b -> bool) -> ... -> ('a,'b,...) typ -> bool

Inherited attribute' is the same as argument, synthetized attribute is GT.comparison.

val trait_name : string

module Make : functor (AstHelpers : GTCommon.GTHELPERS_sig.S) -> sig ... end

val register : unit -> unit