Module GTCommon

module Expander : sig ... end

Expander: base module for Generic Tranformers that utilize plugins.

module GTHELPERS_sig : sig ... end

Signature of AST building functions that are required to support a new backend

module HelpersBase : sig ... end

module Naming : sig ... end

module Plugin : sig ... end

A few base classes for plugins with virtual methods to be implemented.

module Plugin_intf : sig ... end

Base class type that all trait should follow.