Module Stateful

Stateful plugin: functors + inherited value to make decisions about how to map values.

Behave the same as Eval trait but can may return modified state.

Inherited attributes' type (both default and for type parameters) is 'env.

Synthetized attributes' type (both default and for type parameters) is 'env * _ t.

For type declaration type ('a,'b,...) typ = ... it will create transformation function with type

('env -> 'a -> 'env * 'a2) -> ('env -> 'b -> 'env * 'b2) -> ... -> 'env -> ('a,'b,...) typ -> 'env * ('a2, 'b2, ...) typ

val trait_name : string

module Make : functor (AstHelpers : GTCommon.GTHELPERS_sig.S) -> sig ... end

val register : unit -> unit