Module View.HashtblC

Viewing hash tables of Viewable types with explicit concatenation function. Set items are ordered in according to their <b>string representations</b>.

Parameters

module C : Concat

module M : Stdlib.Hashtbl.S

module K : Viewable with type t = M.key

module V : Viewable

Signature

type t = V.t M.t

The type.

val toString : t -> string

View function.