Module View.MapC

Viewing maps of Viewable types with explicit concatenation function. Set items are ordered in according to their <b>string representations</b>.

Parameters

C : Concat

M : Stdlib.Map.S

K : Viewable with type t = M.key

V : Viewable

Signature

type t = V.t M.t

The type.

val toString : t -> string

View function.