Package GT 0.5.2

Compare Compare plugin: receive another value as inherited attribute and compare.

Enum

Eq Eq plugin: receive another value as inherited attribute and test for equality.

Eval Eval plugin: functors + inherited value to make decisions about how to map values.

Foldl Foldl plugin: fold all values in a type.

Foldr Foldr plugin: fold all values in a type.

GT

GTCommon

Gfmt Format module: pretty-prints a value to Format.formatter using Format module.

Ghash Hash plugin.

Gmap Gmap plugin (functor).

HTML Viewing values of various types in HTML format.

Html Html module: converts a value to its html represenation (work in progress).

Pa_gt

Ppx

Ppx_all

Show Show plugin: converts value to a string.

Stateful Stateful plugin: functors + inherited value to make decisions about how to map values.

View Viewing values of various types.

Package info

authors

https://github.com/Kakadu

https://github.com/dboulytchev

changes-files

CHANGES

depends

bisect_ppx

camlp5

conf-m4

dune

logger-p5

ocaml

ocamlfind

ocamlgraph

odig

odoc

pa_ppx

ppx_inline_test

ppxlib

homepage

https://github.com/PLTools/GT

issues

https://github.com/PLTools/GT/issues

license

LGPL-2.1-or-later

maintainers

Kakadu@pm.me

readme-files

README.md

repo

git+https://github.com/PLTools/GT.git

version

0.5.2