Module GTCommon.Plugin_intf

type plugin_args = (Ppxlib.longident * Ppxlib.expression) list

A type that stores additional arguments passed to each plugin.

class virtual ['loc, 'exp, 'typ, 'type_arg, 'ctf, 'cf, 'str, 'sign] typ_g : object ... end

Base class type for all plugins.

module Make : functor (AstHelpers : GTHELPERS_sig.S) -> sig ... end

Functor that takes AST construction functions for a specific backend and constructs a plugin object.

module type MAKE = functor (AstHelpers : GTHELPERS_sig.S) -> sig ... end