Module HTML.Anchor

Functor anchor takes one argument which describes type for which values anchors are set and name to distinguish namespaces (if any).

Parameters

module X : sig ... end

Signature

val set : X.t -> unit

Set anchor for value.

val get : X.t -> string

Get anchor value.

val url : X.t -> string

Get url for the value.

val ref : X.t -> viewer -> viewer

Make reference to the value.

module String : sig ... end

Make reference for strings.